refactored ResourceCard, moved types to component file
