--- apps/web/src/components/GoToCommunity/ResourceCard/resourceCardTypes.ts
@@ -1,9 +1,0 @@
-export type Resource = {
-  title: string;
-  description: string;
-  href: string;
-};
-
-export type ResourceCardProps = Resource & {
-  counter: number;
-};
